+++ skoolit/templates/base.html
@@ -8,8 +8,6 @@
 
 <ul>
 	<li><a href="{{url_for('home')}}">Página inicial</a></li>
-	<li><a href="{{url_for('criar')}}">Criar</a></li>
-	<li><a href="{{url_for('listar')}}">Listar</a></li>
 </ul>
 
 {% block content %}
