+++ skoolit/templates/home.html
@@ -6,4 +6,8 @@
 
 <h1>Skoolit</h1>
 
+<ul>
+	<li><a href="{{url_for('usuarios.listar')}}">Listagem de usuários</a></li>
+</ul>
+
 {% endblock %}
