+++ skoolit/usuarios/templates/usuarios/listar.html
@@ -6,9 +6,11 @@
 
 <h1>Listar usuários</h1>
 
+<a href="{{url_for('usuarios.criar')}}">Criar novo usuário</a>
+
 <ul>
 	{% for usuario in usuarios %}
-		<li>{{usuario.email}} | <a href="{{url_for('atualizar', id=usuario.id)}}">Editar</a> | <a href="{{url_for('excluir', id=usuario.id)}}">Excluir</a> |</li>
+		<li>{{usuario.email}} | <a href="{{url_for('usuarios.atualizar', id=usuario.id)}}">Editar</a> | <a href="{{url_for('usuarios.excluir', id=usuario.id)}}">Excluir</a> |</li>
 	{% endfor %}
 </ul>
